[pcs] Flatten pcs module
